haystack, unstructured -> optional (#413)
